Fix integer overflow in gEDA toKiCadDist()
aMils * MILS_TO_IU multiplies two ints. MILS_TO_IU is 254, so
values above about 8.4 million mils overflow signed int which
is undefined behavior. Widen to int64_t before the multiply.

+++ eeschema/sch_io/geda/sch_io_geda.cpp
@@ -1532,7 +1532,7 @@ bool SCH_IO_GEDA::CanReadSchematicFile( const wxString& aFileName ) const
 
 int SCH_IO_GEDA::toKiCadDist( int aMils ) const
 {
-    return aMils * MILS_TO_IU;
+    return static_cast<int>( static_cast<int64_t>( aMils ) * MILS_TO_IU );
 }
